finally add business logic I. AM. SPEED.

+++ order/src/main/java/org/mehmetcc/order/dto/CreateOrderRequest.java
@@ -24,7 +24,7 @@ public class CreateOrderRequest {
 
     @NotNull
     @Positive
-    private final Integer size;
+    private final BigDecimal size;
 
     @NotNull
     @Positive
